+++ src/StageKit.Fallout/StageKitBuild.MacOS.cs
@@ -207,7 +207,7 @@ public partial class StageKitBuild
         }
     }
 
-    /// <summary>Creates one compressed macOS disk image.</summary>
+    /// <summary>Creates one read-only macOS disk image.</summary>
     protected virtual void CreateMacOSDmg(PublishRidContext context)
     {
         CreateMacOSPackage(context, ApplicationPackagingType.MacOSDmg, ".dmg",
@@ -223,7 +223,7 @@ public partial class StageKitBuild
                 SoftwareVersion, outputPath));
     }
 
-    /// <summary>Creates one multi-architecture compressed macOS disk image.</summary>
+    /// <summary>Creates one multi-architecture read-only macOS disk image.</summary>
     protected virtual void CreateMultiArchMacOSDmg(PublishRidContext x64Context,
         PublishRidContext arm64Context)
     {